Refactor #37 구글맵 마커를 만드는 로직을 utils로 분리
